@@ -1,6 +1,6 @@
 package com.pavel.a692group;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 /**
  * Created by Pavel Mazhnik on 02.03.19.
